Compatibility Improvements
Fixed compatibility issues introduced by CustomNPCs-1.20.1-GBPort-Unofficial-1.20.1.20260711 that prevented the addon from working correctly. Although the official addon is maintained separately from this project, we've added compatibility for worlds created with older versions of the official addon.

+++ src/main/java/com/goodbird/cnpcgeckoaddon/entity/EntityCustomModel.java
@@ -48,7 +48,7 @@ public class EntityCustomModel extends Animal implements GeoAnimatable, GeoEntit
     public ItemStack leftHeldItem;
     public String headBoneName = "head";
     private EntityDimensions dims;
-    public int size = 5;
+    public float size = 5.0f;
     public EntityNPCInterface owner;
     public static final int MAX_ATTACKS = 20;
     public String[] attackAnimNames = new String[MAX_ATTACKS];
